requirements.txt
numpy>=1.24.3
requests>=2.31.0
Pillow>=10.3.0
rawpy>=0.19.1
joblib>=1.3.2
